Add option to augment data with negatives
An additional option is provided to specify a path to a YOLO format
dataset. The items in the dataset are cropped and cached in RAM. During
training each image is augmented by pasting the items such that they
don't occlude existing objects. Additionally, before pasting the
negatives, they are augmented by flipping and resizing.