Feature AFC 1463 | Negative augmentation (#12)
* Add option to augment data with negatives

An additional option is provided to specify a path to a YOLO format
dataset. The items in the dataset are cropped and cached in RAM. During
training each image is augmented by pasting the items such that they
don't occlude existing objects. Additionally, before pasting the
negatives, they are augmented by flipping and resizing.

* Log train batch visualizations to mlflow

* Change default dataset root to directory containing the dataset spec

* Store mlflow run id

Co-authored-by: Jonas Otten <jootten@uos.de>